Move Installer to instrumentation
