+++ agent/agent-core/src/main/java/org/bithon/agent/core/starter/AgentStarter.java
@@ -19,10 +19,10 @@ package org.bithon.agent.core.starter;
 import org.bithon.agent.AgentBuildVersion;
 import org.bithon.agent.core.config.ConfigurationManager;
 import org.bithon.agent.core.interceptor.AopConfig;
-import org.bithon.agent.core.interceptor.installer.InterceptorInstaller;
 import org.bithon.agent.core.interceptor.plugin.PluginResolver;
 import org.bithon.agent.instrumentation.aop.InstrumentationHelper;
 import org.bithon.agent.instrumentation.aop.debug.AopDebugger;
+import org.bithon.agent.instrumentation.aop.interceptor.installer.InterceptorInstaller;
 import org.bithon.agent.instrumentation.loader.AgentClassLoader;
 import org.bithon.agent.instrumentation.utils.AgentDirectory;
 import org.bithon.component.commons.logging.ILogAdaptor;
