+++ agent/agent-plugins/grpc/src/main/java/org/bithon/agent/plugin/grpc/client/interceptor/AbstractAsyncStub$NewStub.java
@@ -19,11 +19,11 @@ package org.bithon.agent.plugin.grpc.client.interceptor;
 import io.grpc.CallOptions;
 import io.grpc.Channel;
 import io.grpc.stub.AbstractStub;
-import org.bithon.agent.core.interceptor.installer.DynamicInterceptorInstaller;
 import org.bithon.agent.instrumentation.aop.advice.AdviceClassGenerator;
 import org.bithon.agent.instrumentation.aop.advice.DynamicAopAdvice;
 import org.bithon.agent.instrumentation.aop.context.AopContext;
 import org.bithon.agent.instrumentation.aop.interceptor.BeforeInterceptor;
+import org.bithon.agent.instrumentation.aop.interceptor.installer.DynamicInterceptorInstaller;
 import org.bithon.agent.instrumentation.aop.interceptor.matcher.Matchers;
 import org.bithon.shaded.net.bytebuddy.asm.Advice;
 import org.bithon.shaded.net.bytebuddy.description.modifier.Visibility;
